requirements_live.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==3.2.23
django-ckeditor==6.3.2
django-js-asset==2.0.0
frozenlist==1.3.3
idna==3.6
importlib-metadata==6.7.0
install==1.3.5
instamojo-wrapper==1.2.0
multidict==6.0.4
mysqlclient==2.1.1
openai==0.28.0
paypalrestsdk==1.5.0
Pillow==8.0.0
PyJWT==2.8.0
PySocks==1.7.1
pytz==2023.3.post1
razorpay==1.4.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sslcommerz-lib==1.0
stripe==7.10.0
tqdm==4.66.1
twilio==6.10.0
typing_extensions==4.7.1
urllib3==1.25.11
whitenoise==6.4.0
yarl==1.9.4
zipp==3.15.0